hoc and navbar changed, styles ordered and general review
