+++ react/tic-tac-toe/src/app/components/Navbar/styles.module.scss
@@ -37,6 +37,7 @@
   font-size: 20px;
   font-weight: 500;
   margin: 0 25px;
+  text-transform: capitalize;
 
   &:hover {
     color: $dark-gray;
