--- react/tic-tac-toe/src/app/screens/Matches/components/MatchesList/styles.module.scss
@@ -1,17 +1,0 @@
-@import 'src/scss/colors.scss';
-@import 'src/scss/fonts.scss';
-
-.item-info {
-  align-items: center;
-  border-bottom: 2px solid $background-color;
-  display: flex;
-  margin: 18px 0;
-  padding: 5px 10px;
-}
-
-.text-info {
-  font-size: $text-detail;
-  margin: 4px 14px;
-  text-align: center;
-  width: 100px;
-}
