+++ react/tic-tac-toe/src/app/screens/Matches/styles.module.scss
@@ -1,14 +1,6 @@
 @import 'src/scss/colors.scss';
 @import 'src/scss/fonts.scss';
 
-.item-info {
-  align-items: center;
-  border-bottom: 2px solid $background-color;
-  display: flex;
-  margin: 18px 0;
-  padding: 5px 10px;
-}
-
 .info-container {
   display: flex;
   min-height: calc(100vh - 70px);
@@ -17,6 +9,9 @@
 }
 
 .list-matches {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
   height: calc(100vh - 70px);
   overflow: scroll;
   padding: 0;
@@ -29,13 +24,6 @@
   display: flex;
   justify-content: space-around;
   width: 70%;
-}
-
-.text-info {
-  font-size: $text-detail;
-  margin: 4px 14px;
-  text-align: center;
-  width: 100px;
 }
 
 .winner-one,
