+++ react/tic-tac-toe/src/scss/application.scss
@@ -19,3 +19,7 @@ code {
 ol, ul {
   padding-left: 30px;
 }
+
+.spinner {
+  margin-top: 30px;
+}
